+++ src/main/java/net/rptools/maptool/client/AppActions.java
@@ -76,6 +76,7 @@ import net.rptools.maptool.model.Zone.Layer;
 import net.rptools.maptool.model.Zone.VisionType;
 import net.rptools.maptool.model.campaign.CampaignManager;
 import net.rptools.maptool.model.drawing.DrawableTexturePaint;
+import net.rptools.maptool.model.grid.Grid;
 import net.rptools.maptool.model.player.*;
 import net.rptools.maptool.model.player.Player.Role;
 import net.rptools.maptool.model.player.PlayerDatabaseFactory.PlayerDatabaseType;
