+++ src/main/java/net/rptools/maptool/client/AppPreferences.java
@@ -28,9 +28,9 @@ import java.util.prefs.Preferences;
 import net.rptools.maptool.client.ui.theme.RessourceManager;
 import net.rptools.maptool.client.walker.WalkerMetric;
 import net.rptools.maptool.language.I18N;
-import net.rptools.maptool.model.GridFactory;
 import net.rptools.maptool.model.Token;
 import net.rptools.maptool.model.Zone;
+import net.rptools.maptool.model.grid.GridFactory;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
