+++ src/main/java/net/rptools/maptool/client/tool/Tool.java
@@ -37,7 +37,7 @@ import javax.swing.KeyStroke;
 import net.rptools.maptool.client.MapTool;
 import net.rptools.maptool.client.ui.zone.ZoneRenderer;
 import net.rptools.maptool.language.I18N;
-import net.rptools.maptool.model.Grid;
+import net.rptools.maptool.model.grid.Grid;
 
 public abstract class Tool extends JToggleButton implements ActionListener, KeyListener {
