+++ src/main/java/net/rptools/maptool/client/ui/zone/ZoneRenderer.java
@@ -70,6 +70,9 @@ import net.rptools.maptool.model.Token.TerrainModifierOperation;
 import net.rptools.maptool.model.Token.TokenShape;
 import net.rptools.maptool.model.Zone.Layer;
 import net.rptools.maptool.model.drawing.*;
+import net.rptools.maptool.model.grid.Grid;
+import net.rptools.maptool.model.grid.GridCapabilities;
+import net.rptools.maptool.model.grid.IsometricGrid;
 import net.rptools.maptool.model.player.Player;
 import net.rptools.maptool.model.zones.*;
 import net.rptools.maptool.util.GraphicsUtil;
