+++ src/main/java/net/rptools/maptool/model/grid/GridCapabilities.java
@@ -12,7 +12,7 @@
  * <http://www.gnu.org/licenses/> and specifically the Affero license
  * text at <http://www.gnu.org/licenses/agpl.html>.
  */
-package net.rptools.maptool.model;
+package net.rptools.maptool.model.grid;
 
 public interface GridCapabilities {
   /**
