+++ src/test/java/net/rptools/maptool/model/drawing/RadiusTemplateAreaTest.java
@@ -20,6 +20,8 @@ import java.awt.*;
 import java.awt.Rectangle;
 import java.awt.geom.Area;
 import net.rptools.maptool.model.*;
+import net.rptools.maptool.model.grid.Grid;
+import net.rptools.maptool.model.grid.SquareGrid;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
